+++ src/main/java/kitchenpos/product/dao/ProductRepository.java
@@ -1,0 +1,18 @@
+package kitchenpos.product.dao;
+
+import kitchenpos.product.domain.Product;
+
+import java.util.List;
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProductRepository extends JpaRepository<Product, Long> {
+    Product save(Product entity);
+
+    Optional<Product> findById(Long id);
+
+    List<Product> findAll();
+}
